feat(data): hanfts v2.0 적용 및 검색 인덱싱 식별자 체계 변경
- 한국어 검색 라이브러리인 `hanfts`를 1.0.0에서 2.0.0으로 업데이트함.
- 검색 엔진의 문서 식별자 단위를 SQLite의 `rowid`(Long)에서 도메인
식별자인 `id`(String)로 변경하여 데이터베이스와 인덱스 간의 참조 구조를
단순화함.
- `SyncDataSourceImpl`에서 데이터를 동기화(upsert/insert)할 때 검색
인덱싱을 함께 수행하도록 개선하여 동기화된 데이터의 검색 가용성을 확보함.
- `LifeLogRepository`에서 로그 삭제(`deleteLifeLogById`) 및 전체
삭제(`clearAllLifeLogs`) 시 검색 인덱스에서도 해당 데이터를 제거하도록
동기화 로직을 추가함.
- `README.md`에 `auth`, `sync`, `common` 등 신규 추가된 모듈 구조를
반영하고, Firebase 및 인증 관련 기술 스택 설명을 업데이트함.
- 식별자 타입 변경에 따라 `FakeSearchEngine`, `TestLifeLogDao` 및 관련
테스트 케이스를 현행화함.

+++ core/data/src/main/java/com/bossmg/android/data/initializer/FtsIndexInitializer.kt
@@ -37,7 +37,7 @@ class FtsIndexInitializer @Inject constructor(
             runCatching {
                 val docs =
                     readRepository.getLifeLogs().first()
-                        .map { Document(id = it.id.toLong(), title = it.title, body = it.description) }
+                        .map { Document(id = it.id, title = it.title, body = it.description) }
                 searchEngine.rebuildIndex(docs)
             }
         }
